# LibJS: Improve garbage collection trigger condition