@awesomekling 2023-08-09
Yeah this seems like an obviously better heuristic. Let's do it! :^)